Add support for 64-bit Windows build: base and chrome/app
This is required to support Native Client on 64-bit Windows.
Native Client will use a small 64-bit executable to load NaCl modules on 64-bit Windows. This 64-bit executable will use Chrome code, but some functionality has to be removed to minimize dependencies. This functionality may be enabled later if we decide that it is necessary.

TEST=None
BUG=28176
Review URL: http://codereview.chromium.org/371073

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@33823 0039d316-1c4b-4281-b951-d872f2087c98

+++ base/file_util_win.cc
@@ -805,8 +805,10 @@ bool MemoryMappedFile::MapFileToMemoryInternal() {
   if (length_ == INVALID_FILE_SIZE)
     return false;
 
+  // length_ value comes from GetFileSize() above. GetFileSize() returns DWORD,
+  // therefore the cast here is safe.
   file_mapping_ = ::CreateFileMapping(file_, NULL, PAGE_READONLY,
-                                      0, length_, NULL);
+                                      0, static_cast<DWORD>(length_), NULL);
   if (file_mapping_ == INVALID_HANDLE_VALUE)
     return false;
